# _workspace/02_backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: KAMOS API
  version: 0.1.0
  summary: KAMOS — Japanese alcoholic beverage tracking and discovery (MVP).
  description: |
    REST API for the KAMOS mobile app. All responses are JSON.

    **i18n contract:** beverage / brewery / category / flavor-tag names are
    returned as `{ "en": "...", "ja": "...", "ko": "..." }` objects. Per SPEC
    §6.5, clients fall back to `en` when the requested locale is missing.
    The server does NOT pre-resolve names based on `Accept-Language`.

    **Pagination:** every list endpoint returns the canonical envelope
    `{ items, next_cursor, has_more }`. The cursor is an opaque base64 string
    encoding the last item's keyset. Never offset.

    **Error shape:** `{ "error": "<human>", "code": "<MACHINE_CODE>" }` for
    every non-2xx response.

    **Auth:** Bearer JWT (HS256). Public endpoints are explicitly tagged. Some
    public endpoints accept optional auth to surface follow state.
servers:
  - url: http://localhost:8080
    description: local dev
...
security:
  - bearerAuth: []
paths:
...
  /v1/beverages:
    get:
      tags: [beverages]
      operationId: listBeverages
      security: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string, enum: [nihonshu, shochu, liqueur] }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: page of beverages, sorted by popularity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageOfBeverage" }
  /v1/beverages/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [beverages]
      operationId: getBeverage
      security: []
      responses:
        "200":
          description: beverage detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BeverageDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/beverages/{id}/check-ins:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: cursor
        schema: { type: string }
      - in: query
        name: limit
        schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
    get:
      tags: [beverages, checkins]
      operationId: getBeverageCheckins
      security: []
      responses:
        "200":
          description: page of check-ins for this beverage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageOfCheckinSummary" }
  /v1/breweries:
    get:
      tags: [breweries]
      operationId: listBreweries
      security: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: page of breweries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageOfBrewery" }
  /v1/breweries/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [breweries]
      operationId: getBrewery
      security: []
      responses:
        "200":
          description: brewery detail with beverage list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Brewery"
                  - type: object
                    properties:
                      beverages: { $ref: "#/components/schemas/PageOfBeverage" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ----- CHECK-INS -----
...
  /v1/collections:
    get:
      tags: [collections]
      operationId: listCollections
      responses:
        "200":
          description: list of own collections
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageOfCollection" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [collections]
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 50 }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /v1/collections/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: cursor
        schema: { type: string }
      - in: query
        name: limit
        schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    get:
      tags: [collections]
      operationId: getCollection
      responses:
        "200":
          description: collection with entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Collection"
                  - type: object
                    properties:
                      entries: { $ref: "#/components/schemas/PageOfCollectionEntry" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [collections]
      operationId: renameCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 50 }
      responses:
        "200":
          description: renamed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [collections]
      operationId: deleteCollection
      responses:
        "204": { description: soft-deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/collections/{id}/entries:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [collections]
      operationId: addCollectionEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [beverage_id]
              properties:
                beverage_id: { type: string, format: uuid }
                note: { type: string, maxLength: 200, nullable: true }
      responses:
        "204": { description: added (idempotent) }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/collections/{id}/entries/{beverage_id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: beverage_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [collections]
      operationId: updateCollectionEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 200, nullable: true }
      responses:
        "204": { description: updated }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [collections]
      operationId: removeCollectionEntry
      responses:
        "204": { description: removed }
        "404": { $ref: "#/components/responses/NotFound" }

  # ----- SEARCH -----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: missing or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: not found (or hidden by privacy)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: uniqueness or state conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Validation:
      description: validation failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: human-readable message
        code:
          type: string
          description: machine code (UPPER_SNAKE_CASE)
    I18nText:
      type: object
      description: |
        i18n object. `en` and `ja` are typically present; `ko` is optional.
        Per SPEC §6.5 fallback, clients should display `ko` if present, else
        `en`. Same rule for ja → en.
      required: [en]
      properties:
        en: { type: string }
        ja: { type: string }
        ko: { type: string }
    CategoryLabel:
      type: object
      required: [slug, label_i18n]
      properties:
        slug:
          type: string
          enum: [nihonshu, shochu, liqueur]
        label_i18n:
          $ref: "#/components/schemas/I18nText"
          description: |
            SPEC §2.1 canonical strings:
            * en: 'Nihonshu (Sake)' | 'Shochu' | 'Liqueur'
            * ja: '日本酒' | '焼酎' | 'リキュール'
            * ko: '니혼슈 (사케)' | '쇼츄' | '리큐어'
...
    Brewery:
      type: object
      required: [id, name, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { $ref: "#/components/schemas/I18nText" }
        prefecture: { type: string, nullable: true }
        region: { type: string, nullable: true }
        founded_year: { type: integer, nullable: true }
        website: { type: string, format: uri, nullable: true }
        description: { $ref: "#/components/schemas/I18nText", nullable: true }
        created_at: { type: string, format: date-time }
    Beverage:
      type: object
      required: [id, name, brewery, category, flavor_profile, check_in_count, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { $ref: "#/components/schemas/I18nText" }
        brewery: { $ref: "#/components/schemas/Brewery" }
        category: { $ref: "#/components/schemas/CategoryLabel" }
        subcategory: { $ref: "#/components/schemas/I18nText", nullable: true }
        abv: { type: number, nullable: true }
        polishing_ratio: { type: integer, nullable: true, description: "Nihonshu only" }
        prefecture: { type: string, nullable: true }
        region: { type: string, nullable: true }
        flavor_profile:
          type: array
          items: { type: string, description: "flavor tag slug" }
        description: { $ref: "#/components/schemas/I18nText", nullable: true }
        label_image_url: { type: string, format: uri, nullable: true }
        avg_rating: { type: number, nullable: true, minimum: 0.5, maximum: 5.0 }
        check_in_count: { type: integer }
        created_at: { type: string, format: date-time }
    FlavorAggregate:
      type: object
      required: [slug, dimension, name, uses]
      properties:
        slug: { type: string }
        dimension: { type: string, enum: [sweetness, body, acidity, character, finish] }
        name: { $ref: "#/components/schemas/I18nText" }
        uses: { type: integer }
...
    BeverageDetail:
      allOf:
        - $ref: "#/components/schemas/Beverage"
        - type: object
          required: [aggregated_flavor, recent_check_ins]
          properties:
            aggregated_flavor:
              type: array
              items: { $ref: "#/components/schemas/FlavorAggregate" }
            recent_check_ins:
              type: array
              items: { $ref: "#/components/schemas/CheckinSummary" }
    CheckinUser:
      type: object
      required: [id, username, display_username, display_name]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        display_username: { type: string }
        display_name: { type: string }
        avatar_url: { type: string, format: uri, nullable: true }
    BeverageRef:
      type: object
      required: [id, name, brewery, category]
      properties:
        id: { type: string, format: uuid }
        name: { $ref: "#/components/schemas/I18nText" }
        brewery: { $ref: "#/components/schemas/BreweryRef" }
        category: { $ref: "#/components/schemas/CategoryLabel" }
        label_image_url: { type: string, format: uri, nullable: true }
    BreweryRef:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { $ref: "#/components/schemas/I18nText" }
        region: { type: string, nullable: true }
...
    CheckinSummary:
      type: object
      required: [id, user, created_at]
      properties:
        id: { type: string, format: uuid }
        user: { $ref: "#/components/schemas/CheckinUser" }
        rating: { type: number, nullable: true, minimum: 0.5, maximum: 5.0 }
        review: { type: string, nullable: true, maxLength: 500 }
        created_at: { type: string, format: date-time }
...
    Collection:
      type: object
      required: [id, name, entry_count, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 50 }
        entry_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CollectionEntry:
      type: object
      required: [beverage, added_at]
      properties:
        beverage: { $ref: "#/components/schemas/BeverageRef" }
        note: { type: string, maxLength: 200, nullable: true }
        added_at: { type: string, format: date-time }
...
    PageBase:
      type: object
      required: [items, has_more]
      properties:
        next_cursor:
          type: string
          description: "opaque keyset cursor; absent on last page"
        has_more: { type: boolean }
    PageOfBeverage:
      allOf:
        - $ref: "#/components/schemas/PageBase"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Beverage" }
    PageOfBrewery:
      allOf:
        - $ref: "#/components/schemas/PageBase"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Brewery" }
...
    PageOfCheckinSummary:
      allOf:
        - $ref: "#/components/schemas/PageBase"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/CheckinSummary" }
...
    PageOfCollection:
      allOf:
        - $ref: "#/components/schemas/PageBase"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Collection" }
    PageOfCollectionEntry:
      allOf:
        - $ref: "#/components/schemas/PageBase"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/CollectionEntry" }
...